feat: extract logs routes to routes/logs.py
