+++ templates/_error_panels.html
@@ -254,7 +254,7 @@
                         <div class="d-flex gap-2 mt-2">
                             {% set _trace_ids_csv = err.trace_ids_csv if err.trace_ids_csv is defined else '' %}
                             {% if _trace_ids_csv %}
-                                <a href="{{ url_for('view_logs', trace_ids=_trace_ids_csv, from_ts=from_ts, to_ts=to_ts) }}"
+                                <a href="{{ url_for('logs.view_logs', trace_ids=_trace_ids_csv, from_ts=from_ts, to_ts=to_ts) }}"
                                    class="btn btn-sm btn-outline-{{ logs_btn_outline }}">
                                     <i class="bi bi-journal-text me-1"></i><span class="error-panel-btn-label">{{ logs_label }}</span>
                                 </a>
@@ -265,7 +265,7 @@
                                     </a>
                                 {% endif %}
                             {% elif err.trace_id %}
-                                <a href="{{ url_for('view_logs', trace_id=err.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
+                                <a href="{{ url_for('logs.view_logs', trace_id=err.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
                                    class="btn btn-sm btn-outline-{{ logs_btn_outline }}">
                                     <i class="bi bi-journal-text me-1"></i><span class="error-panel-btn-label">{{ logs_label }}</span>
                                 </a>
@@ -274,7 +274,7 @@
                                     <i class="bi bi-diagram-3 me-1"></i><span class="error-panel-btn-label">{{ trace_label }}</span>
                                 </a>
                             {% else %}
-                                <a href="{{ url_for('view_logs', service=err.service, from_ts=from_ts, to_ts=to_ts) }}"
+                                <a href="{{ url_for('logs.view_logs', service=err.service, from_ts=from_ts, to_ts=to_ts) }}"
                                    class="btn btn-sm btn-outline-{{ logs_btn_outline }}">
                                     <i class="bi bi-journal-text me-1"></i><span class="error-panel-btn-label">{{ logs_label }}</span>
                                 </a>
