+++ templates/base.html
@@ -730,7 +730,7 @@
         </a>
       </li>
       <li class="nav-item">
-        <a href="{{ url_for('view_logs') }}" class="nav-link {% if request.endpoint == 'view_logs' %}active{% endif %}" title="Logs">
+        <a href="{{ url_for('logs.view_logs') }}" class="nav-link {% if request.endpoint == 'logs.view_logs' %}active{% endif %}" title="Logs">
           <i class="bi bi-journal-text"></i><span class="sb-label">Logs</span>
         </a>
       </li>
