+++ templates/custom_dashboard_view.html
@@ -578,7 +578,7 @@
   var templatesMap = {};
   TEMPLATES.forEach(function (t) { templatesMap[t.id] = t; });
   var sourceUrls = {
-    logs: {{ url_for('view_logs') | tojson }},
+    logs: {{ url_for('logs.view_logs') | tojson }},
     metrics: {{ url_for('view_metrics_anomaly') | tojson }},
     traces: {{ url_for('view_traces') | tojson }},
     errors: {{ url_for('errors.view_errors') | tojson }}
