+++ templates/logs_help.html
@@ -10,7 +10,7 @@
     <h4 class="fw-semibold mb-1">Logs Help</h4>
     <p class="text-secondary mb-0">Real-time log streaming, regex search, advanced analysis, and live mode.</p>
   </div>
-  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('view_logs') }}" title="Back to Logs">
+  <a class="btn btn-outline-secondary btn-sm" href="{{ url_for('logs.view_logs') }}" title="Back to Logs">
     <i class="bi bi-arrow-left me-1"></i><span class="help-btn-label">Back to Logs</span>
   </a>
 </div>
