+++ templates/metrics_anomaly.html
@@ -27,7 +27,7 @@
        data-ai-risk="low"
        data-ai-confirm="false"
        title="Open Related Logs"
-       href="{{ url_for('view_logs', service=service, from_ts=from_ts, to_ts=to_ts, stats='1') }}">
+       href="{{ url_for('logs.view_logs', service=service, from_ts=from_ts, to_ts=to_ts, stats='1') }}">
       <i class="bi bi-journal-text me-1"></i><span class="metrics-anomaly-btn-label">Open Related Logs</span>
     </a>
     {% elif related_target == 'traces' %}
@@ -223,7 +223,7 @@
             <td data-label="Actions">
               {% if r.related_target == 'logs' %}
               <a class="btn btn-sm btn-outline-info"
-                 href="{{ url_for('view_logs', service=r.service, from_ts=r.time, window_s=60, stats='1') }}">
+                 href="{{ url_for('logs.view_logs', service=r.service, from_ts=r.time, window_s=60, stats='1') }}">
                 Logs
               </a>
               {% elif r.related_target == 'traces' %}
