+++ templates/traces.html
@@ -644,7 +644,7 @@
 
       <!-- Logs link -->
       {% if log_count > 0 %}
-      <a href="{{ url_for('view_logs', trace_id=span.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
+      <a href="{{ url_for('logs.view_logs', trace_id=span.trace_id, from_ts=from_ts, to_ts=to_ts) }}"
          class="btn btn-link btn-sm p-0 ms-1 text-info" style="font-size:.75rem; line-height:1;"
          title="{{ log_count }} associated log record(s)">
         <i class="bi bi-journal-text"></i><span class="ms-1">{{ log_count }}</span>
